Enabling user to define macros for all compiled files.
Checks for the existence of buildprefs.txt in the sketch directory. If
it exists then any settings for compiler.c.extra_flags and
compiler.cpp.extra_flags override earlier definitions. With the -D
option to gcc/g++ the user can define macros which are used for all
compilation parts, including libraries.